+++ components/src/androidUnitTest/snapshots/images/com.alexrdclement.uiplayground.components_MediaControlSheetTest_mediaControlSheet[PartiallyExpanded].png
