+++ components/src/androidUnitTest/snapshots/images/com.alexrdclement.uiplayground.components_PlayPauseButtonTest_playPauseButton[isPlaying=false,isEnabled=true].png
