+++ components/src/commonMain/kotlin/com/alexrdclement/uiplayground/components/Checkbox.kt
@@ -2,17 +2,11 @@ package com.alexrdclement.uiplayground.components
 
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.semantics.Role
 import androidx.compose.ui.semantics.role
 import androidx.compose.ui.semantics.semantics
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.tooling.preview.PreviewParameter
-import com.alexrdclement.uiplayground.components.preview.BoolPreviewParameterProvider
 import com.alexrdclement.uiplayground.theme.PlaygroundTheme
 
 @Composable
@@ -36,22 +30,5 @@ fun Checkbox(
             text = if (isChecked) "☑︎" else "☐",
             style = PlaygroundTheme.typography.titleLarge
         )
-    }
-}
-
-@Preview
-@Composable
-private fun Preview(
-    @PreviewParameter(BoolPreviewParameterProvider::class) isChecked: Boolean,
-) {
-    PlaygroundTheme {
-        Surface {
-            var isChecked by remember { mutableStateOf(isChecked) }
-
-            Checkbox(
-                isChecked = isChecked,
-                onCheckedChange = { isChecked = !it },
-            )
-        }
     }
 }
