+++ components/src/commonMain/kotlin/com/alexrdclement/uiplayground/components/MediaControlSheet.kt
@@ -11,36 +11,24 @@ import androidx.compose.foundation.gestures.Orientation
 import androidx.compose.foundation.gestures.anchoredDraggable
 import androidx.compose.foundation.gestures.animateTo
 import androidx.compose.foundation.gestures.snapTo
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.BoxWithConstraints
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.navigationBarsPadding
 import androidx.compose.foundation.layout.offset
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.Stable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.saveable.Saver
 import androidx.compose.runtime.saveable.rememberSaveable
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.onSizeChanged
 import androidx.compose.ui.platform.LocalDensity
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Density
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.dp
-import androidx.tracing.trace
-import com.alexrdclement.uiplayground.components.model.Artist
+import androidx.compose.ui.util.trace
 import com.alexrdclement.uiplayground.components.model.MediaItem
-import com.alexrdclement.uiplayground.theme.PlaygroundTheme
-import kotlinx.coroutines.launch
 import kotlin.math.max
 
 private const val TraceName = "MediaControlSheet"
@@ -233,41 +221,4 @@ private fun Modifier.modalBottomSheetAnchors(
     }
 
     state.anchoredDraggableState.updateAnchors(newAnchors, newTarget)
-}
-
-@Preview(showBackground = true)
-@Composable
-private fun Preview() {
-    PlaygroundTheme {
-        val mediaItem = MediaItem(
-            artworkThumbnailUrl = null,
-            artworkLargeUrl = null,
-            title = "Title",
-            artists = listOf(Artist("Artist 1"), Artist("Artist 2")),
-        )
-        var isPlaying by remember { mutableStateOf(false) }
-        val state = rememberMediaControlSheetState(
-            initialValue = MediaControlSheetAnchor.PartiallyExpanded,
-        )
-        val coroutineScope = rememberCoroutineScope()
-        MediaControlSheet(
-            mediaItem = mediaItem,
-            isPlaying = isPlaying,
-            onPlayPauseClick = { isPlaying = !isPlaying },
-            onControlBarClick = {
-                coroutineScope.launch {
-                    state.expand()
-                }
-            },
-            state = state,
-        ) {
-            Column(
-                verticalArrangement = Arrangement.Center,
-                horizontalAlignment = Alignment.CenterHorizontally,
-                modifier = Modifier.fillMaxSize(),
-            ) {
-                Text("Content")
-            }
-        }
-    }
 }
