+++ components/src/commonMain/kotlin/com/alexrdclement/uiplayground/components/PlayPauseButton.kt
@@ -6,15 +6,12 @@ import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Pause
+import androidx.compose.material.icons.filled.Close
 import androidx.compose.material.icons.filled.PlayArrow
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.ColorFilter
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.unit.dp
-import com.alexrdclement.uiplayground.components.preview.BoolPreviewParameterProvider
 import com.alexrdclement.uiplayground.theme.PlaygroundTheme
 
 @Composable
@@ -33,41 +30,13 @@ fun PlayPauseButton(
             .aspectRatio(1f)
     ) {
         Image(
-            imageVector = if (isPlaying) Icons.Default.Pause else Icons.Default.PlayArrow,
+            imageVector = if (isPlaying) Icons.Default.Close else Icons.Default.PlayArrow,
             contentDescription = if (isPlaying) {
                 PlayPauseButtonContentDescriptionPlaying
             } else PlayPauseButtonContentDescriptionPaused,
             colorFilter = ColorFilter.tint(PlaygroundTheme.colorScheme.onPrimary),
             modifier = Modifier
                 .fillMaxSize()
-        )
-    }
-}
-
-@Preview
-@Composable
-private fun PlayPreview(
-    @PreviewParameter(BoolPreviewParameterProvider::class) isEnabled: Boolean,
-) {
-    PlaygroundTheme {
-        PlayPauseButton(
-            isEnabled = isEnabled,
-            isPlaying = false,
-            onClick = {},
-        )
-    }
-}
-
-@Preview
-@Composable
-private fun PausePreview(
-    @PreviewParameter(BoolPreviewParameterProvider::class) isEnabled: Boolean,
-) {
-    PlaygroundTheme {
-        PlayPauseButton(
-            isEnabled = isEnabled,
-            isPlaying = true,
-            onClick = {},
         )
     }
 }
