+++ components/src/commonMain/kotlin/com/alexrdclement/uiplayground/components/Slider.kt
@@ -39,7 +39,6 @@ import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.semantics.disabled
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.semantics.setProgress
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.util.lerp
 import com.alexrdclement.uiplayground.theme.PlaygroundTheme
@@ -362,14 +361,3 @@ private fun scale(a1: Float, b1: Float, x1: Float, a2: Float, b2: Float) =
 // Calculate the 0..1 fraction that `pos` value represents between `a` and `b`
 private fun calcFraction(a: Float, b: Float, pos: Float) =
     (if (b - a == 0f) 0f else (pos - a) / (b - a)).coerceIn(0f, 1f)
-
-@Preview
-@Composable
-private fun Preview() {
-    PlaygroundTheme {
-        Surface {
-            val state = remember { SliderState(value = 0.5f) }
-            Slider(state = state)
-        }
-    }
-}
