+++ components/src/commonMain/kotlin/com/alexrdclement/uiplayground/components/Surface.kt
@@ -19,9 +19,6 @@ import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.semantics.isTraversalGroup
 import androidx.compose.ui.semantics.semantics
-import androidx.compose.ui.tooling.preview.PreviewLightDark
-import androidx.compose.ui.tooling.preview.PreviewParameter
-import com.alexrdclement.uiplayground.components.preview.BoolPreviewParameterProvider
 import com.alexrdclement.uiplayground.theme.PlaygroundTheme
 
 @Composable
@@ -99,29 +96,3 @@ private fun Modifier.surface(
     .then(if (border != null) Modifier.border(border, shape) else Modifier)
     .background(color = backgroundColor, shape = shape)
     .clip(shape)
-
-
-@PreviewLightDark
-@Composable
-private fun Preview() {
-    PlaygroundTheme {
-        Surface {
-            Text("Hello world")
-        }
-    }
-}
-
-@PreviewLightDark
-@Composable
-private fun PreviewClickable(
-    @PreviewParameter(BoolPreviewParameterProvider::class) enabled: Boolean,
-) {
-    PlaygroundTheme {
-        Surface(
-            onClick = {},
-            enabled = enabled
-        ) {
-            Text("Hello world")
-        }
-    }
-}
