+++ components/src/commonMain/kotlin/com/alexrdclement/uiplayground/components/model/Artist.kt
