+++ gradle/libs.versions.toml
@@ -8,7 +8,7 @@ androidx-test-ext-junit = "1.2.1"
 androidx-tracing = "1.3.0-alpha02"
 baselineprofile = "1.3.0"
 benchmark-macro-junit4 = "1.3.0"
-coil = "2.7.0"
+coil = "3.0.0-rc01"
 compose-bom = "2024.08.00"
 compose-multiplatform = "1.7.0"
 compose-runtime-tracing = "1.0.0-beta01"
@@ -17,9 +17,8 @@ core-ktx = "1.13.1"
 espresso-core = "3.6.1"
 junit = "4.13.2"
 kotlin = "2.0.21"
+ktor = "3.0.0"
 lifecycle-runtime-ktx = "2.8.4"
-material3 = "1.2.1"
-material3-icons-extended = "1.6.8"
 paparazzi = "1.3.4"
 profileinstaller = "1.3.1"
 test-parameter-injector = "1.14"
@@ -38,8 +37,8 @@ androidx-tracing = { group = "androidx.tracing", name = "tracing-ktx", version.r
 androidx-tracing-perfetto = { group = "androidx.tracing", name = "tracing-perfetto", version.ref = "tracing-perfetto" }
 androidx-tracing-perfetto-binary = { group = "androidx.tracing", name = "tracing-perfetto-binary", version.ref = "tracing-perfetto" }
 androidx-uiautomator = { group = "androidx.test.uiautomator", name = "uiautomator", version.ref = "uiautomator" }
-coil = { group = "io.coil-kt", name = "coil", version.ref = "coil" }
-coil-compose = { group = "io.coil-kt", name = "coil-compose", version.ref = "coil" }
+coil-compose = { group = "io.coil-kt.coil3", name = "coil-compose", version.ref = "coil" }
+coil-network-ktor3 = { group = "io.coil-kt.coil3", name = "coil-network-ktor3", version.ref = "coil" }
 core-ktx = { group = "androidx.core", name = "core-ktx", version.ref = "core-ktx" }
 compose-bom = { group = "androidx.compose", name = "compose-bom", version.ref = "compose-bom" }
 compose-foundation = { group = "androidx.compose.foundation", name = "foundation" }
@@ -52,9 +51,10 @@ compose-ui-test-manifest = { group = "androidx.compose.ui", name = "ui-test-mani
 compose-ui-test-junit4 = { group = "androidx.compose.ui", name = "ui-test-junit4", version.ref = "compose-ui" }
 espresso-core = { group = "androidx.test.espresso", name = "espresso-core", version.ref = "espresso-core" }
 junit = { group = "junit", name = "junit", version.ref = "junit" }
+ktor-client-android = { group = "io.ktor", name = "ktor-client-android", version.ref = "ktor" }
+ktor-client-darwin = { group = "io.ktor", name = "ktor-client-darwin", version.ref = "ktor" }
+ktor-client-java = { group = "io.ktor", name = "ktor-client-java", version.ref = "ktor" }
 lifecycle-runtime-ktx = { group = "androidx.lifecycle", name = "lifecycle-runtime-ktx", version.ref = "lifecycle-runtime-ktx" }
-material3 = { group = "androidx.compose.material3", name = "material3", version.ref = "material3" }
-material-icons-extended = { group = "androidx.compose.material", name = "material-icons-extended", version.ref = "material3-icons-extended" }
 paparazzi = { group = "app.cash.paparazzi", name = "paparazzi", version.ref = "paparazzi" }
 test-parameter-injector = { group = "com.google.testparameterinjector", name = "test-parameter-injector", version.ref = "test-parameter-injector" }
